+++ test/integration/open_citations/test_open_citations_datasource.py
@@ -4,6 +4,7 @@
 
 from unittest import TestCase
 
+from tol.core import DataSourceFilter
 from tol.sources.open_citations import open_citations
 
 
@@ -30,3 +31,33 @@ class TestOpenCitationsDataSource(TestCase):
         self.assertIsNone(next(ret))
         with self.assertRaises(StopIteration):
             next(ret)
+
+    def test_get_list_by_doi_and_pmid(self):
+        ods = open_citations()
+
+        objects = list(ods.get_list(
+            'meta',
+            DataSourceFilter(and_={
+                'reference_id': {
+                    'in_list': {
+                        'value': [
+                            '10.1038/nphys1170',
+                            'pmid:23287718',
+                        ],
+                    },
+                },
+            }),
+        ))
+
+        self.assertEqual(
+            {obj.id for obj in objects},
+            {'10.1038/nphys1170', '10.1126/science.1231143'},
+        )
+        objects_by_doi = {obj.id: obj for obj in objects}
+        pmid_object = objects_by_doi['10.1126/science.1231143']
+        self.assertEqual(pmid_object.pmid, '23287718')
+        self.assertEqual(
+            pmid_object.title,
+            'Multiplex Genome Engineering Using CRISPR/Cas Systems',
+        )
+        self.assertEqual(pmid_object.pub_date, '2013-01-03')